[Feature 160840449] restructure auth file
